+++ src/apis/movie/useGetNowPlayingMovies.ts
@@ -6,24 +6,7 @@ import {
 } from '../../types/NowPlayingMovies';
 import { DataQueryKeys } from '../data-query-keys';
 import { endPoints } from '../endPoints';
-
-const snakeCaseToCamelCase = (s: string) => {
-  return s.replace(/([-_][a-z])/gi, $1 => {
-    return $1.toUpperCase().replace('-', '').replace('_', '');
-  });
-};
-
-const convertKeysToCamelCase = (obj: any): any => {
-  if (Array.isArray(obj)) {
-    return obj.map(v => convertKeysToCamelCase(v));
-  } else if (obj !== null && typeof obj === 'object') {
-    return Object.keys(obj).reduce((acc: Record<string, any>, key) => {
-      acc[snakeCaseToCamelCase(key)] = convertKeysToCamelCase(obj[key]);
-      return acc;
-    }, {});
-  }
-  return obj;
-};
+import { convertKeysToCamelCase } from '../../components/ConvertKeysToCamelCase';
 
 export const useGetNowPlayingMovie = () => {
   return useQuery<NowPlayingMovies, Error>({
